% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{add_all_composites}
\alias{add_all_composites}
\title{Add every composite defined in a composite-definitions table}
\usage{
add_all_composites(scores_df, composite_defs, suffix = "_ssGSEA")
}
\arguments{
\item{scores_df}{Data frame of \verb{<Signature>_ssGSEA} columns.}

\item{composite_defs}{Table from \code{\link[=load_composite_defs]{load_composite_defs()}}.}

\item{suffix}{Score column suffix.}
}
\description{
Add every composite defined in a composite-definitions table
}
