% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{detect_id_type}
\alias{detect_id_type}
\title{Heuristically detect a gene-ID vector's type}
\usage{
detect_id_type(genes)
}
\arguments{
\item{genes}{Character vector of gene IDs.}
}
\value{
\code{"ensembl"} if more than 80\% of non-NA entries start with \code{ENSG},
else \code{"symbol"}.
}
\description{
Heuristically detect a gene-ID vector's type
}
