% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{gene_set_coverage}
\alias{gene_set_coverage}
\title{Percentage of a gene set present in an expression matrix}
\usage{
gene_set_coverage(mat, genes)
}
\arguments{
\item{mat}{An expression matrix with gene IDs as row names.}

\item{genes}{Character vector of gene IDs.}
}
\description{
Percentage of a gene set present in an expression matrix
}
