% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{metric_cols}
\alias{metric_cols}
\title{Get survival time and event column names for a metric}
\usage{
metric_cols(metric)
}
\arguments{
\item{metric}{"OS" or "RFS".}
}
\value{
Named character vector with elements \code{t} and \code{e}.
}
\description{
Get survival time and event column names for a metric
}
