% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pub_palette}
\alias{pub_palette}
\title{Standardized color palette for outcomes, groups, and sample types}
\usage{
pub_palette
}
\description{
Standardized color palette for outcomes, groups, and sample types
}
