% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversion.R
\name{strip_ensembl_version}
\alias{strip_ensembl_version}
\title{Strip an Ensembl gene-ID version suffix}
\usage{
strip_ensembl_version(ids)
}
\arguments{
\item{ids}{Character vector of Ensembl IDs.}
}
\description{
\code{ENSG00000123456.3} -> \code{ENSG00000123456}.
}
