% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{tcga_patient_id}
\alias{tcga_patient_id}
\title{Extract the 12-character TCGA patient ID from a full barcode}
\usage{
tcga_patient_id(barcodes)
}
\arguments{
\item{barcodes}{Character vector of TCGA barcodes/aliquot IDs.}
}
\description{
Extract the 12-character TCGA patient ID from a full barcode
}
